add separation of structs in a new file, delete duplicate Mocks and add comments to methods
